python_application/requirements.txt
boto3==1.22.2
botocore==1.25.2
Flask==2.0.3
gunicorn==20.1.0
Jinja2==3.0.3
jmespath==0.10.0
jsonify==0.5
MarkupSafe==2.0.1
psycopg2==2.9.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
s3transfer==0.5.0
num2words~=0.5.10
